Move CommandUsageMessage to its own file
