+++ bot/action/core/command/__init__.py
@@ -1,7 +1,7 @@
 from bot.action.core.action import IntermediateAction
 from bot.action.core.command.parser import UnderscoredCommandParser, CommandParser
 from bot.action.core.command.throttler.shortlyrepeatedcommand import ShortlyRepeatedCommandThrottler
-from bot.api.domain import MessageEntityParser, Message
+from bot.api.domain import MessageEntityParser
 
 
 class CommandAction(IntermediateAction):
@@ -33,27 +33,6 @@ class CommandAction(IntermediateAction):
     @staticmethod
     def is_valid_command(entity):
         return entity.type == "bot_command" and entity.offset == 0
-
-
-class CommandUsageMessage:
-    @classmethod
-    def get_usage_message(cls, command, args=None, description=""):
-        message = "*Usage*\n"
-        if type(args) is list:
-            message += "\n".join((cls.__get_command_with_args(command, arg) for arg in args))
-        else:
-            message += cls.__get_command_with_args(command, args)
-        if description:
-            message += "\n\n" + description
-        return Message.create(message, parse_mode="Markdown", disable_web_page_preview=True)
-
-    @staticmethod
-    def __get_command_with_args(command, args):
-        text = "`" + command
-        if args:
-            text += " " + args
-        text += "`"
-        return text
 
 
 class UnderscoredCommandBuilder:
