+++ bot/action/extra/audios.py
@@ -2,7 +2,8 @@ import collections
 
 from bot.action import util
 from bot.action.core.action import Action
-from bot.action.core.command import UnderscoredCommandBuilder, CommandUsageMessage
+from bot.action.core.command import UnderscoredCommandBuilder
+from bot.action.core.command.usagemessage import CommandUsageMessage
 from bot.action.userinfo import UserStorageHandler
 from bot.action.util.format import UserFormatter, DateFormatter, TimeFormatter, SizeFormatter
 from bot.action.util.textformat import FormattedText
