+++ bot/action/extra/messages.py
@@ -3,7 +3,8 @@ import json
 
 from bot.action.chatsettings import ChatSettings
 from bot.action.core.action import Action
-from bot.action.core.command import CommandUsageMessage, UnderscoredCommandBuilder
+from bot.action.core.command import UnderscoredCommandBuilder
+from bot.action.core.command.usagemessage import CommandUsageMessage
 from bot.action.userinfo import UserStorageHandler
 from bot.action.util.format import UserFormatter, DateFormatter
 from bot.action.util.textformat import FormattedText
