+++ bot/action/extra/pole.py
@@ -4,7 +4,8 @@ import datetime
 import pytz
 
 from bot.action.core.action import Action
-from bot.action.core.command import CommandUsageMessage, UnderscoredCommandBuilder
+from bot.action.core.command import UnderscoredCommandBuilder
+from bot.action.core.command.usagemessage import CommandUsageMessage
 from bot.action.userinfo import UserStorageHandler
 from bot.action.util.format import DateFormatter, UserFormatter
 from bot.action.util.textformat import FormattedText
